+++ docs/changes.rst
@@ -4,6 +4,8 @@ Changelog
 6.0.2 (Unreleased)
 ------------------
 
+- Add page number to RTF reports. This fixes 
+  `TNO ticket 241 <https://code.simplon.biz/tracker/tno-euphorie/ticket/241>`_.
 - Fixed #7261: Typo in splash page
 - Added Bulgarian help headers, per #7317
 - Portuguese translations of the splash page. #7870.